# avaliacao-api/src/main/resources/webapp/files/openapi.yaml
openapi: 3.0.1
info:
  title: API de estados 
  description: avaliação 3 | Jhonatan Alves Pereira | CompassUOL
  version: 0.0.1
  termsOfService: https://github/JhonatanAlves4/
servers: 
-   url: http://localhost:8080
paths:
  /api/states:
    get:
      summary: Recupera todos os estados filtrados pela maior população e maior area.
      parameters: 
      - name: regiao
        in: query
        schema:
          type: string
      responses:
        200: 
          description: Sucesso.
          content:
            application/json: 
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    id: 
                      type: integer
                    nome: 
                      type: string
                    regiao:
                      type: string
                    populacao:
                      type: integer
                    capital:
                      type: string
                    area: 
                      type: integer
    post:
      summary: Adiciona um novo estado.
      description: "Atenção ao adicionar o atributo de região, apenas serão aceitos os valores: NORTE, SUL, CENTRO_OESTE, SUDESTE e NORDESTE."
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nome: 
                  type: string
                regiao:
                  type: string
                populacao:
                  type: integer
                capital:
                  type: string
                area: 
                  type: integer
      responses:
        201: 
          description: Criado com sucesso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  nome: 
                    type: string
                  regiao:
                    type: string
                  populacao:
                    type: integer
                  capital:
                    type: string
                  area: 
                    type: integer
          
  /api/states/{id}:
    get:
      summary: Recupera o estado do respectivo id.
      parameters: 
      - name: id
        in: path
        schema:
          type: integer
        required: true
      responses:
        200: 
          description: Sucesso
          content:
            application/json: 
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    id: 
                      type: integer
                    nome: 
                      type: string
                    regiao:
                      type: string
                    populacao:
                      type: integer
                    capital:
                      type: string
                    area: 
                      type: integer
        400:
          description: Erro, estado não encontrado
          content:
            application/json: 
              schema:
                type: object
                properties:
                  message: 
                    type: string
                  httpStatus: 
                    type: string
                  timestamp:
                    type: string
    put:
      summary: Altera o estado do respectivo id.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nome: 
                  type: string
                regiao:
                  type: string
                populacao:
                  type: integer
                capital:
                  type: string
                area: 
                  type: integer
      parameters: 
      - name: id
        in: path
        schema:
          type: integer
        required: true
      responses:
        200: 
          description: Sucesso
          content:
            application/json: 
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    id: 
                      type: integer
                    nome: 
                      type: string
                    regiao:
                      type: string
                    populacao:
                      type: integer
                    capital:
                      type: string
                    area: 
                      type: integer
        400:
          description: Erro, estado não encontrado
          content:
            application/json: 
              schema:
                type: object
                properties:
                  message: 
                    type: string
                  httpStatus: 
                    type: string
                  timestamp:
                    type: string
    delete:
      summary: Deleta o estado do respectivo id.
      parameters: 
      - name: id
        in: path
        schema:
          type: integer
        required: true
      responses:
        200: 
          description: Sucesso